% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{callgama}
\alias{callgama}
\title{Call GAMA}
\usage{
callgama(experiment, output = "output", Xms = 512, Xmx = 2048)
}
\arguments{
\item{experiment}{Either an R experiment object or the path to an XML file
describing the experiment.}

\item{output}{Name of the output directory.}

\item{Xms}{Initial memory allocation pool for the Java Virtual Machine.}

\item{Xmx}{Maximal memory allocation pool for the Java Virtual Machine.}
}
\description{
Call the simualation platform GAMA to run an experiment described in an XML
file.
}
\keyword{internal}

