% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{checkdir}
\alias{checkdir}
\title{Check a directory}
\usage{
checkdir(dir)
}
\arguments{
\item{dir}{Name of the directory to check and optionally rename.}
}
\description{
Checks the existance of the directory "dir". If it exists, it renames it.
}
\keyword{internal}

