% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{readxml}
\alias{readxml}
\title{Reads a GAMA output file}
\usage{
readxml(file)
}
\arguments{
\item{file}{Name of the XML file to read.}
}
\description{
Reads a GAMA output in an XML file and return in the form of a datarame with
one column per "monitor".
}
\keyword{internal}

