% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run a GAMA experiment}
\usage{
run(experiment, Xms = 512, Xmx = 2048)
}
\arguments{
\item{experiment}{Either an R experiment object or the path to an XML file
describing the experiment.}

\item{Xms}{Initial memory allocation pool for the Java Virtual Machine.}

\item{Xmx}{Maximal memory allocation pool for the Java Virtual Machine.}
}
\value{
A data frame with, for each column, the values of the model state
variables for each time step.
}
\description{
Run a GAMA experiment from an XML file.
}
\details{
If _experiment_ is an XML, it should contains the path to a GAML
file that describes the model. A GAML file starts with a model name and
should contain three sections defining (1) the world (global), (2) the
species (species), and (3) the experiments (experiment). The XML file should
contain a Simulation header and two slots: (1) Parameters et (2) Outputs. See
the _demo_ folder of the _gamar_ library for examples of such files. See
\url{https://github.com/gama-platform/gama/wiki/G__Headless} for more
detailed information on how these XML and GAML files should be structured.
}
\examples{
## Running an SIR epidemic model.
# Copying the example files to the working directory:
file.copy(system.file("examples/sir.xml",package="gamar"),".")
file.copy(system.file("examples/sir.gaml",package="gamar"),".")
# Running the experiment:
simulation <- run("sir.xml")
# Removing the example files from the working directory:
file.remove("sir.xml","sir.gaml")
# Plotting the simulations output:
with(simulation, {
  plot(Susceptible,col="blue",type="l",xlab="time",ylab="number of individuals")
  lines(Infectious,col="red")
  lines(Recovered,col="green")
})
# Adding the legend:
legend("right",c("susceptible","infected","recovered"),
  col=c("blue","red","green"),lty=1,bty="n")
}
\author{
Marc Choisy, Jean-Daniel Zucker
}

